save an image
wrote a function that takes in a query and saves an image to your
current directory of the image that is returned by that query. Also
realized I have been switching the lat and lon and that explained why I
was not able to find some images. I retrieved an image of UCB campus
taken last week.